requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
python-jose[cryptography]==3.3.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.2
openai==1.45.0
Pillow==10.4.0
python-dateutil==2.9.0
bson==0.5.10